melhoria 1 no controller

+++ Controllers/HomeController.cs
@@ -10,51 +10,61 @@ namespace Todo.Controllers;
 
         [HttpGet("/")]
         // [Route("/")]
-        public List<TodoModel> Get([FromServices]AppDbContext context){ //metodos dentro de um controller são actions
-            return context.Todos.ToList();
-        }
+        public IActionResult Get([FromServices]AppDbContext context) //metodos dentro de um controller são actions
+           =>  Ok(context.Todos.ToList());
+        
            
      [HttpGet("/{id:int}")]
         // [Route("/")]
-        public TodoModel GetById([FromRoute] int id
-            ,[FromServices]AppDbContext context){
+        public IActionResult GetById([FromRoute] int id
+            ,[FromServices]AppDbContext context)
+            {
+        var todos = context.Todos.FirstOrDefault(x => x.Id == id);
 
-            return context.Todos.FirstOrDefault(x => x.Id == id);
+        if(todos == null)
+        return NotFound();
+        
+        return Ok(todos);
+
         }
+
 
 
         [HttpPost("/")]
-        public TodoModel Post([FromBody]TodoModel todo,
+        public IActionResult Post([FromBody]TodoModel todo,
             [FromServices] AppDbContext context){
+
                 context.Todos.Add(todo);
                 context.SaveChanges();
-                return todo;
+
+                return Created($"/{todo.Id}", todo);
         }
 
              [HttpPut("/{id:int}")]
-        public TodoModel Put(
+        public IActionResult Put(
             [FromRoute] int id,
             [FromBody]TodoModel todo,
             [FromServices] AppDbContext context){
                 var model = context.Todos.FirstOrDefault(x => x.Id == id);
                 if(model == null )
-                    return todo;
+                    return NotFound();
                
                 model.Title = todo.Title;
                 model.Done = todo.Done;
                 context.Todos.Update(model);
                 context.SaveChanges();
-                return todo;
+                return Ok(model);
         }
         [HttpDelete("/{id:int}")]
-        public TodoModel Delete(
+        public IActionResult Delete(
             [FromRoute] int id,
             [FromServices] AppDbContext context){
                 var model = context.Todos.FirstOrDefault(x => x.Id == id);
-
+                if(model == null )
+                    return NotFound();
                 context.Todos.Remove(model);
                 context.SaveChanges();
-                return model;
+                return Ok(model);
         }
